Use utf8mb4 to connect to the database.
if you use a mysql client and do "SET NAMES utf8mb4;" the pm and
tag names should be correct now. Test this by creating a pm and or
tag with unicode characters like ë, æ, ç or 😉

This will *NOT* correct previously created data. Convert this data
before switching or you will get mojibake.